davinci: da850/omap-l138 evm: account for defdcdc{2,3} being tied high
Per the da850/omap-l138 Beta EVM SOM schematic, the defdcdc2 and
defdcdc3 lines are tied high by default. This leads to a 3.3V IO
and 1.2V CVDD voltage.

For the TPS6507x driver, it means that the DEFDCDC_HIGH registers
need to be read/written instead of DEFDCDC_LOW register. The tps6507x
driver added this functionality recently. This patch passes the right
platform data to the TPS6507x which tells it that the HIGH register
needs to be read instead of the LOW register.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>