davinci: da830/omap-l137: Configure LCD_AC_ENB_CS pin as GPIO
On AM17xx EVMs, GLCD will not work unless the GPIO1[11]
pin is driven high. This pin was configured to be a LCD
functional pin earlier. Configuring this pin as GPIO will
not have any effect on the older DA830/OMP-L137 EVMs.

Also, LCD backlight pin needs to be driven high on AM17xx
EVMs.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>
Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>